pyproject.toml
[tool.poetry]
name = "llm-ol"
version = "0.1.0"
description = ""
authors = ["Andy Lo <andylolu24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.1"
requests = "^2.31.0"
absl-py = "^2.0.0"
aiohttp = "^3.9.0"
arxiv-base = "^0.17.8"
urllib3 = "<2"             # Resolution hint
tqdm = "^4.66.1"
networkx = "^3.2.1"
matplotlib = "^3.8.2"
scipy = "^1.11.4"
pygraphviz = "^1.11"
pandas = "^2.1.3"
datasets = "^2.15.0"
kaggle = "^1.5.16"
python-dotenv = "^1.0.0"
notebook = "^7.0.7"
transformers = {extras = ["torch"], version = "^4.36.2"}
ipywidgets = "^8.1.1"
scikit-learn = "^1.4.0"
seaborn = "^0.13.1"
spacy = "^3.7.2"
nltk = "^3.8.1"
torch-geometric = "^2.4.0"
huggingface-hub = "^0.20.3"
jinja2 = "^3.1.3"
outlines = "0.0.34"
trl = "^0.7.11"
scienceplots = "^2.1.1"
openai = "^1.13.3"
vllm = "^0.3.3"
ml-collections = "^0.1.1"
wandb = "^0.16.4"
tensorboardx = "^2.6.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_third_party = ["wandb"]